test: the fork-only families build under clang with warnings as errors
clang 22 rejects ninety-three constructs in these units that GCC accepts, all of
them in the tests but one:

  - forty-nine of the reflection macro's hidden friend left unemitted, because
    the block it belongs to sits in an anonymous namespace and nothing in the
    unit calls the friend. The nineteen units now put their helper blocks in a
    named namespace with a using-directive after it, the form the other tests
    use.
  - twenty locals whose names boost::ut also declares in the scope the case runs
    in: nine RuntimeTest instances named "test", seven sample counts named
    "skip", four tag loop variables named "tag".
  - nineteen float values widened to double inside an expression, eighteen of
    them the real and imaginary parts handed to a std::complex<double>, now
    stated with a cast at the point of widening.
  - three shifts and complements whose operands are int against unsigned, one of
    them in p25/ImbeFrame.hpp, where imbeUnpackUVector() accumulates the
    parameter word in unsigned as the packing side beside it already does.
  - one aggregate initialized with braces where the refusal case needs only the
    settings, and one lambda capture of a constant the body reads by value.

The behavior of every case is unchanged: no assertion, no bound and no fixture
moves. The audio test is upstream's, unmodified here, and its six missing-field
initializers are left with it.

+++ blocks/p25/test/qa_P25FrameLayer.cpp
@@ -167,7 +167,7 @@ const boost::ut::suite<"P25FrameLayer"> p25FrameLayerTests = [] {
             const std::size_t at     = duidTransmittedDibits(kDuids[0]);
             for (unsigned i = 0U; i < 12U; ++i) {
                 const std::size_t j = at + kFrameSyncDibits + i;
-                stream[j]           = static_cast<std::uint8_t>(~stream[j] & 0x3U);
+                stream[j]           = static_cast<std::uint8_t>(~static_cast<unsigned>(stream[j]) & 0x3U);
             }
             const Caught c        = run(stream);
             bool         reported = false;
